Create SusyUtility.h (declarations) and SusyUtility.cc (definitions).
